# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Kumquat API
  version: 1.0.0
  description: Northbound business API. Engine custom resources are internal execution projections.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /roles/{id}:
    parameters: [{$ref: '#/components/parameters/NumericID'}]
    get: {operationId: getRole, responses: {'200': {$ref: '#/components/responses/Role'}, '404': {$ref: '#/components/responses/Error'}}}
...
  /projects:
    get:
      operationId: listProjects
      parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/Size'}]
      responses: {'200': {$ref: '#/components/responses/ProjectList'}}
    post:
      operationId: createProject
      requestBody: {$ref: '#/components/requestBodies/ProjectCreate'}
      responses: {'200': {$ref: '#/components/responses/Project'}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /projects/{id}:
    parameters: [{$ref: '#/components/parameters/PublicID'}]
    get: {operationId: getProject, responses: {'200': {$ref: '#/components/responses/Project'}, '404': {$ref: '#/components/responses/Error'}}}
    put:
      operationId: updateProject
      requestBody: {$ref: '#/components/requestBodies/ProjectUpdate'}
      responses: {'200': {$ref: '#/components/responses/Project'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
    delete: {operationId: deleteProject, responses: {'200': {$ref: '#/components/responses/Message'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
...
  /clusters/{id}:
    parameters: [{$ref: '#/components/parameters/PublicID'}]
    get: {operationId: getCluster, responses: {'200': {$ref: '#/components/responses/Resource'}, '404': {$ref: '#/components/responses/Error'}}}
  /clusters/{id}/adopt:
    parameters: [{$ref: '#/components/parameters/PublicID'}]
    post:
      operationId: adoptCluster
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'202': {$ref: '#/components/responses/Accepted'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /operations/{id}:
    parameters: [{$ref: '#/components/parameters/PublicID'}]
    get: {operationId: getOperation, responses: {'200': {$ref: '#/components/responses/Operation'}, '404': {$ref: '#/components/responses/Error'}}}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    PublicID: {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/PublicID'}}
    NumericID: {name: id, in: path, required: true, schema: {type: integer, format: int64, minimum: 1}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: false, schema: {type: string, minLength: 1, maxLength: 128}}
    Page: {name: page, in: query, required: false, schema: {type: integer, minimum: 1, default: 1}}
    Size: {name: size, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
  requestBodies:
...
    ProjectCreate: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProjectCreate'}}}}
    ProjectUpdate: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProjectUpdate'}}}}
...
  schemas:
    PublicID: {type: string, minLength: 3, maxLength: 40, pattern: '^[a-z]+_[a-f0-9]+$'}
...
    ProjectConfig: {type: object, additionalProperties: {oneOf: [{type: string}, {type: number}, {type: integer}, {type: boolean}]}}
    ProjectCreate: {type: object, additionalProperties: false, required: [name, moduleId], properties: {name: {type: string}, moduleId: {$ref: '#/components/schemas/PublicID'}, config: {$ref: '#/components/schemas/ProjectConfig'}}}
    ProjectUpdate: {type: object, additionalProperties: false, properties: {name: {type: string}, moduleId: {$ref: '#/components/schemas/PublicID'}, config: {$ref: '#/components/schemas/ProjectConfig'}}}
    WorkloadRef: {type: object, additionalProperties: false, required: [apiVersion, kind], properties: {apiVersion: {type: string}, kind: {type: string, enum: [Deployment, StatefulSet, DaemonSet, ReplicaSet, Job, CronJob]}}}
...
    EnvironmentVariable: {type: object, additionalProperties: false, required: [name, valueFrom], properties: {name: {type: string}, valueFrom: {$ref: '#/components/schemas/SecretKeyRef'}}}
    PolicyOverride: {type: object, additionalProperties: false, properties: {clusterSelector: {$ref: '#/components/schemas/LabelSelector'}, image: {type: string}, env: {type: array, items: {$ref: '#/components/schemas/EnvironmentVariable'}}, resources: {type: object}, command: {type: array, items: {type: string}}, args: {type: array, items: {type: string}}}}
    Container: {type: object, additionalProperties: false, required: [name, image], properties: {name: {type: string}, image: {type: string}, command: {type: array, items: {type: string}}, args: {type: array, items: {type: string}}, env: {type: array, items: {$ref: '#/components/schemas/EnvironmentVariable'}}}}
    PodTemplate: {type: object, additionalProperties: false, required: [containers], properties: {labels: {type: object, additionalProperties: {type: string}}, containers: {type: array, minItems: 1, items: {$ref: '#/components/schemas/Container'}}}}
    LabelSelectorRequirement: {type: object, additionalProperties: false, required: [key, operator], properties: {key: {type: string}, operator: {type: string, enum: [In, NotIn, Exists, DoesNotExist]}, values: {type: array, items: {type: string}}}}
    LabelSelector: {type: object, additionalProperties: false, properties: {matchLabels: {type: object, additionalProperties: {type: string}}, matchExpressions: {type: array, items: {$ref: '#/components/schemas/LabelSelectorRequirement'}}}}
    JobAttributes: {type: object, additionalProperties: false, properties: {completions: {type: integer, format: int32}, parallelism: {type: integer, format: int32}, backoffLimit: {type: integer, format: int32}, ttlSecondsAfterFinished: {type: integer, format: int32}, successfulJobsHistoryLimit: {type: integer, format: int32}, failedJobsHistoryLimit: {type: integer, format: int32}}}
    Toleration: {type: object, additionalProperties: false, properties: {key: {type: string}, operator: {type: string}, value: {type: string}, effect: {type: string}, tolerationSeconds: {type: integer, format: int64}}}
    ResiliencyPolicy: {type: object, additionalProperties: false, properties: {minAvailable: {oneOf: [{type: integer}, {type: string}]}, maxUnavailable: {oneOf: [{type: integer}, {type: string}]}}}
    ApplicationDesired: {type: object, additionalProperties: false, required: [workload, template], properties: {workload: {$ref: '#/components/schemas/WorkloadRef'}, selector: {$ref: '#/components/schemas/LabelSelector'}, replicas: {type: integer, format: int32, minimum: 0}, template: {$ref: '#/components/schemas/PodTemplate'}, schedule: {type: string}, suspend: {type: boolean}, jobAttributes: {$ref: '#/components/schemas/JobAttributes'}, clusterAffinity: {type: object}, clusterTolerations: {type: array, items: {$ref: '#/components/schemas/Toleration'}}, overrides: {type: array, items: {$ref: '#/components/schemas/PolicyOverride'}}, resiliency: {$ref: '#/components/schemas/ResiliencyPolicy'}, rolloutStrategy: {type: object}}}
    WorkspaceDesired: {type: object, additionalProperties: false, properties: {namespace: {type: string}, clusterMatchLabels: {type: object, additionalProperties: {type: string}}, clusterSelector: {$ref: '#/components/schemas/LabelSelector'}, resourceConstraints: {type: object}}}
    WorkspaceDesiredState: {type: object, additionalProperties: false, required: [workspace], properties: {workspace: {$ref: '#/components/schemas/WorkspaceDesired'}}}
    ApplicationDesiredState: {type: object, additionalProperties: false, required: [application], properties: {application: {$ref: '#/components/schemas/ApplicationDesired'}}}
...
    ClusterApplicationStatus: {type: object, additionalProperties: false, required: [clusterName, phase, replicas, readyReplicas], properties: {clusterName: {type: string}, phase: {type: string}, replicas: {type: integer, format: int32}, readyReplicas: {type: integer, format: int32}, message: {type: string}}}
    ApplicationObserved: {type: object, additionalProperties: false, required: [schedulingPhase, healthPhase, globalReplicas, globalReadyReplicas], properties: {schedulingPhase: {type: string}, healthPhase: {type: string}, globalReplicas: {type: integer, format: int32}, globalReadyReplicas: {type: integer, format: int32}, clusters: {type: array, items: {$ref: '#/components/schemas/ClusterApplicationStatus'}}}}
    WorkspaceClusterError: {type: object, additionalProperties: false, required: [name, message, lastTransitionTime], properties: {name: {type: string}, message: {type: string}, lastTransitionTime: {type: string, format: date-time}}}
    WorkspaceObserved: {type: object, additionalProperties: false, required: [ready], properties: {ready: {type: boolean}, appliedClusters: {type: array, items: {type: string}}, failedClusters: {type: array, items: {$ref: '#/components/schemas/WorkspaceClusterError'}}}}
    ClusterObserved: {type: object, additionalProperties: false, required: [connectionMode, state], properties: {connectionMode: {type: string, enum: [Edge]}, state: {type: string}, kubernetesVersion: {type: string}, agentVersion: {type: string}, lastKeepAliveTime: {type: string, format: date-time, nullable: true}}}
    ObservedState: {type: object, additionalProperties: false, properties: {application: {$ref: '#/components/schemas/ApplicationObserved'}, workspace: {$ref: '#/components/schemas/WorkspaceObserved'}, cluster: {$ref: '#/components/schemas/ClusterObserved'}}}
    EmptyDesiredState: {type: object, additionalProperties: false, maxProperties: 0}
    ResourceBase: {type: object, required: [id, kind, name, observed, syncState, source], properties: {id: {$ref: '#/components/schemas/PublicID'}, kind: {type: string, enum: [workspace, application, cluster]}, name: {type: string}, parentId: {$ref: '#/components/schemas/PublicID'}, projectId: {$ref: '#/components/schemas/PublicID'}, moduleId: {$ref: '#/components/schemas/PublicID'}, observed: {$ref: '#/components/schemas/ObservedState'}, observedAt: {type: string, format: date-time}, syncState: {type: string, enum: [pending, retrying, synced, failed, deleting, stale]}, source: {type: string, enum: [api, agent, engine]}, state: {type: string, enum: [discovered, adopting, adopted, archived, unassigned]}}}
    WorkspaceResource: {allOf: [{$ref: '#/components/schemas/ResourceBase'}, {type: object, required: [desired], properties: {kind: {type: string, enum: [workspace]}, desired: {$ref: '#/components/schemas/WorkspaceDesiredState'}}}]}
    ApplicationResource: {allOf: [{$ref: '#/components/schemas/ResourceBase'}, {type: object, required: [desired], properties: {kind: {type: string, enum: [application]}, desired: {$ref: '#/components/schemas/ApplicationDesiredState'}}}]}
    ClusterResource: {allOf: [{$ref: '#/components/schemas/ResourceBase'}, {type: object, required: [desired], properties: {kind: {type: string, enum: [cluster]}, desired: {$ref: '#/components/schemas/EmptyDesiredState'}}}]}
    Resource: {oneOf: [{$ref: '#/components/schemas/WorkspaceResource'}, {$ref: '#/components/schemas/ApplicationResource'}, {$ref: '#/components/schemas/ClusterResource'}], discriminator: {propertyName: kind}}
    Operation: {type: object, additionalProperties: false, required: [id, resourceId, action, state, createdAt, updatedAt], properties: {id: {$ref: '#/components/schemas/PublicID'}, resourceId: {$ref: '#/components/schemas/PublicID'}, action: {type: string, enum: [apply, delete, adopt, adopt-existing]}, state: {type: string, enum: [pending, retrying, succeeded, failed]}, error: {type: string}, createdAt: {type: string, format: date-time}, updatedAt: {type: string, format: date-time}}}
    ErrorEnvelope: {type: object, additionalProperties: false, required: [code, message], properties: {code: {type: integer}, message: {type: string}}}
    RoleDTO: {type: object, additionalProperties: false, required: [id, name, created_at, updated_at], properties: {id: {type: integer, format: int64}, name: {type: string}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}}
    ModuleDTO: {type: object, additionalProperties: false, required: [id, name, level, sort, path, created_at, updated_at], properties: {id: {$ref: '#/components/schemas/PublicID'}, name: {type: string}, parent_id: {$ref: '#/components/schemas/PublicID'}, level: {type: integer}, sort: {type: integer}, path: {type: string}, children: {type: array, items: {$ref: '#/components/schemas/ModuleDTO'}}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}}
...
    ProjectDTO: {type: object, additionalProperties: false, required: [id, name, module_id, created_at, updated_at], properties: {id: {$ref: '#/components/schemas/PublicID'}, name: {type: string}, module_id: {$ref: '#/components/schemas/PublicID'}, module: {$ref: '#/components/schemas/ModuleDTO'}, config: {$ref: '#/components/schemas/ProjectConfig'}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}}
...
    Pagination: {type: object, additionalProperties: false, required: [total, page, size, list], properties: {total: {type: integer, format: int64}, page: {type: integer}, size: {type: integer}, list: {type: array, items: {type: object}}}}
    MessageEnvelope: {type: object, additionalProperties: false, required: [code, message], properties: {code: {type: integer, enum: [0]}, message: {type: string}}}
...
    RoleEnvelope: {type: object, additionalProperties: false, required: [code, message, data], properties: {code: {type: integer, enum: [0]}, message: {type: string}, data: {$ref: '#/components/schemas/RoleDTO'}}}
...
    ProjectEnvelope: {type: object, additionalProperties: false, required: [code, message, data], properties: {code: {type: integer, enum: [0]}, message: {type: string}, data: {$ref: '#/components/schemas/ProjectDTO'}}}
...
    ProjectListEnvelope: {type: object, additionalProperties: false, required: [code, message, data], properties: {code: {type: integer, enum: [0]}, message: {type: string}, data: {allOf: [{$ref: '#/components/schemas/Pagination'}, {type: object, properties: {list: {type: array, items: {$ref: '#/components/schemas/ProjectDTO'}}}}]}}}
...
    ResourceEnvelope: {type: object, additionalProperties: false, required: [code, message, data], properties: {code: {type: integer, enum: [0]}, message: {type: string}, data: {$ref: '#/components/schemas/Resource'}}}
...
    OperationEnvelope: {type: object, additionalProperties: false, required: [code, message, data], properties: {code: {type: integer, enum: [0]}, message: {type: string}, data: {$ref: '#/components/schemas/Operation'}}}
  responses:
...
    Role: {description: Role, content: {application/json: {schema: {$ref: '#/components/schemas/RoleEnvelope'}}}}
...
    Project: {description: Project, content: {application/json: {schema: {$ref: '#/components/schemas/ProjectEnvelope'}}}}
    ProjectList: {description: Projects, content: {application/json: {schema: {$ref: '#/components/schemas/ProjectListEnvelope'}}}}
    Message: {description: Message, content: {application/json: {schema: {$ref: '#/components/schemas/MessageEnvelope'}}}}
    Resource: {description: Resource, content: {application/json: {schema: {$ref: '#/components/schemas/ResourceEnvelope'}}}}
...
    Operation: {description: Operation, content: {application/json: {schema: {$ref: '#/components/schemas/OperationEnvelope'}}}}
    Accepted: {description: Accepted operation, content: {application/json: {schema: {$ref: '#/components/schemas/OperationEnvelope'}}}}
    Error: {description: Error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}